+++ src/main_mapobjects.py
@@ -1,0 +1,4 @@
+from main_triangulate import *
+from visual_slam import *
+
+
