--- src/mapobjects.py
@@ -1,4 +1,0 @@
-from triangulate import *
-from visual_slam import *
-
-
